backend/requirements.txt
aiohttp==3.11.0
APScheduler==3.10.4
fastapi[all]==0.115.8
feedparser==6.0.11
langchain-community==0.3.17
langchain_core==0.3.35
langchain_openai==0.3.5
langchain-anthropic==0.3.5
langchain-google-genai==0.3.5
mitreattack_python==3.0.8
newspaper4k==0.9.3.1
openai==1.62.0
Pillow==11.1.0
pydantic==2.10.6
regex==2022.3.2
Requests==2.32.3
SQLAlchemy==2.0.36
tweepy==4.9.0
uvicorn==0.34.0
lxml[html_clean]